requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
itsdangerous>=2.2.0
PyJWT>=2.8.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
httpx>=0.27.0
requests>=2.31.0
Pillow>=10.0.0
